refactor(a2a_dart): Use optional Logger instead of log level
Instead of passing a log level to the constructors of `A2AClient`, `A2AServer`, and the transports, this change modifies them to accept an optional `Logger` instance.

If a logger is not supplied, then no logging will be performed, which is more efficient than checking the log level on every call. This also simplifies the API by removing the need for `hierarchicalLoggingEnabled`.